GameObject link fixed. Deleting children moved to transform destructor.
